@@ -1,364 +1,574 @@
-"""PyTorch Lightning DataModules for TorchSigNarrowband and TorchSigWideband
-"""
-from torch.utils.data import DataLoader
-from torch.nn import Identity
-import pytorch_lightning as pl
-import numpy as np
-from typing import Callable, Union, Optional
-import os
+"""PyTorch Lightning DataModules for Narrowband and Wideband
 
-from torchsig.datasets.torchsig_narrowband import TorchSigNarrowband
-from torchsig.datasets.torchsig_wideband import TorchSigWideband
-from torchsig.datasets import conf
-from torchsig.utils.dataset import collate_fn as collate_fn_default
-from torchsig.datasets.wideband import WidebandModulationsDataset
-from torchsig.datasets.modulations import ModulationsDataset
+Learn More: https://lightning.ai/docs/pytorch/stable/data/datamodule.html
+
+If dataset does not exist at root, creates new dataset and writes to disk
+If dataset does exsit, simply loaded it back in
+"""
+
+from __future__ import annotations
+
+# TorchSig
+from torch.utils.data import DataLoader
+from torchsig.datasets.dataset_metadata import DatasetMetadata, NarrowbandMetadata, WidebandMetadata
+from torchsig.datasets.narrowband import NewNarrowband, StaticNarrowband
+from torchsig.datasets.wideband import NewWideband, StaticWideband
+from torchsig.datasets.dataset_utils import to_dataset_metadata
 from torchsig.utils.writer import DatasetCreator
-from torchsig.transforms.target_transforms import DescToClassIndex#, DescToListTuple
-from torchsig.transforms.transforms import (
-    Compose,
-    RandomPhaseShift,
-    Normalize,
-    ComplexTo2D,
-    Spectrogram
-)
+from torchsig.utils.file_handlers.base_handler import TorchSigFileHandler
+from torchsig.utils.file_handlers.zarr import ZarrFileHandler
+from torchsig.transforms.base_transforms import Transform
+from torchsig.transforms.target_transforms import TargetTransform
+from torchsig.datasets.default_configs.loader import get_default_yaml_config
+
+# Third Party
+import pytorch_lightning as pl
+
+# Built-In
+from typing import Callable, List
+from pathlib import Path
+
 
 class TorchSigDataModule(pl.LightningDataModule):
-    """General TorchSig DataModule.
-
-        Subclasses must implement `prepare_data`, `setup` functions.
     """
+    PyTorch Lightning DataModule for managing TorchSig datasets.
 
-    def __init__(self, 
-                 root: str, 
-                 dataset: str, 
-                 impaired: bool, 
-                 qa: bool = True, 
-                 eb_no: bool = False, 
-                 seed: int = 12345,
-                 overlap_prob: Optional[float] = .1,
-                 transform: Optional[Callable] = None,
-                 target_transform: Optional[Callable] = None,
-                 batch_size: int = 4,
-                 num_workers: int = 1,
-                 collate_fn: Optional[Callable] = None,
-                ):             
-        """TorchSig DataModule Init
+    Args:
+        root (str): The root directory where datasets are stored or created.
+        dataset (str): The name of the dataset (either 'narrowband' or 'wideband').
+        train_metadata (DatasetMetadata | str | dict): Metadata for the training dataset.
+        val_metadata (DatasetMetadata | str | dict): Metadata for the validation dataset.
+        batch_size (int, optional): The batch size for data loading. Defaults to 1.
+        num_workers (int, optional): The number of worker processes for data loading. Defaults to 1.
+        collate_fn (Callable, optional): A function to collate data into batches.
+        create_batch_size (int, optional): The batch size used during dataset creation. Defaults to 8.
+        create_num_workers (int, optional): The number of workers used during dataset creation. Defaults to 4.
+        file_handler (TorchSigFileHandler, optional): The file handler for managing data storage. Defaults to ZarrFileHandler.
+        transforms (list, optional): A list of transformations to apply to the input data. Defaults to an empty list.
+        target_transforms (list, optional): A list of transformations to apply to the target labels. Defaults to an empty list.
+    """
+    def __init__(
+        self,
+        root: str,
+        dataset: str,
+        train_metadata: DatasetMetadata | str | dict,
+        val_metadata: DatasetMetadata | str | dict,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: list = [],
+        target_transforms: list = [],
+    ):
+        """
+        Initializes the TorchSigDataModule with the provided parameters.
 
         Args:
-            root (str): Dataset root path.
-            dataset (str): Dataset name, either "Narrowband" or "Wideband".
-            impaired (bool): Dataset impairment setting.
-            qa (bool, optional): Generate small dataset sample. Defaults to True.
-            eb_no (bool, optional): Use EbNo config. Defaults to False.
-            seed (int, optional): Dataset generation seed. Defaults to 12345.
-            transform (Optional[Callable], optional): Data transforms. Defaults to None.
-            target_transform (Optional[Callable], optional): Label transforms. Defaults to None.
-            batch_size (int, optional): Dataloader batch size. Defaults to 4.
-            num_workers (int, optional): Dataloader number of workers to use. Defaults to 1.
-            collate_fn (Optional[Callable], optional): Dataloader custom collate function. Defaults to TorchSig collate_fn.
+            root (str): The root directory where datasets are stored or created.
+            dataset (str): The name of the dataset (either 'narrowband' or 'wideband').
+            train_metadata (DatasetMetadata | str | dict): Metadata for the training dataset.
+            val_metadata (DatasetMetadata | str | dict): Metadata for the validation dataset.
+            batch_size (int, optional): The batch size for data loading. Defaults to 1.
+            num_workers (int, optional): The number of worker processes for data loading. Defaults to 1.
+            collate_fn (Callable, optional): A function to collate data into batches.
+            create_batch_size (int, optional): The batch size used during dataset creation. Defaults to 8.
+            create_num_workers (int, optional): The number of workers used during dataset creation. Defaults to 4.
+            file_handler (TorchSigFileHandler, optional): The file handler for managing data storage. Defaults to ZarrFileHandler.
+            transforms (list, optional): A list of transformations to apply to the input data. Defaults to an empty list.
+            target_transforms (list, optional): A list of transformations to apply to the target labels. Defaults to an empty list.
         """
+        # read from yaml or dataset metadata or code inputs
         super().__init__()
 
-        self.root = root
+        self.root = Path(root)
         self.dataset = dataset
-        self.impaired = impaired
-        self.clean = "impaired" if impaired else "clean" #: boolean: `self.impaired` as a string variable.
-        self.seed = seed
-        self.overlap_prob = overlap_prob
+        self.train_metadata = to_dataset_metadata(train_metadata)
+        self.val_metadata = to_dataset_metadata(val_metadata)
+        self.impaired = self.train_metadata.impairment_level > 0
 
-        self.train_config = self._set_config(self.dataset, self.impaired, True, qa, eb_no, self.seed) #: dict: TorshSig train config 
-        self.val_config = self._set_config(self.dataset, self.impaired, False, qa, eb_no, self.seed) #: dict: TorshSig val config 
+        self.new_dataset_class = NewNarrowband if self.dataset == "narrowband" else NewWideband
+        self.static_dataset_class = StaticNarrowband if self.dataset == "narrowband" else StaticWideband
+     
+        self.transforms = transforms
+        self.target_transforms = target_transforms
 
-        self.train = None #: Dataset: train dataset
-        self.val = None #: Dataset: validation dataset
-        
 
-        self.transform = Identity() if transform is None else transform
-        
-        self.target_transform = Identity() if target_transform is None else target_transform
-
+        # initialize dataloader params
         self.batch_size = batch_size
         self.num_workers = num_workers
         self.collate_fn = collate_fn
 
+        # dataset creator params
+        self.create_batch_size = create_batch_size
+        self.create_num_workers = create_num_workers
+        self.file_handler = file_handler
 
-        self.data_path = None #: str: Path to downloaded dataset in root
-        self.train_path = None #: str: Path to train dataset
-        self.val_path = None #: str: Path to validation dataset
-        
-
-    def _set_config(self, dataset: str, impaired: bool, is_train: bool, qa: bool, eb_no: bool, seed: int) -> dict:
-        """_summary_
-
-        Args:
-            dataset (str): Dataset name, either "Narrowband" or "Wideband".
-            impaired (bool): Whether dataset is impaired or not (clean).
-            is_train (bool): Whether dataset is train or not (val).
-            qa (bool): Whether to create smaller dataset version.
-            eb_no (bool): Whether to use EbNo for dataset.
-            seed (int): Seed for dataset generation.
-
-        Raises:
-            ValueError: Dataset name is not Narrowband or Wideband.
-
-        Returns:
-            dict: TorchSig config file for dataset.
-        """
-        if not dataset in ["Narrowband", "Wideband"]:
-            raise ValueError(f"Invalid dataset type: {dataset}")
-        
-        i = "Impaired" if impaired else "Clean"
-        t = "Train" if is_train else "Val"
-        q = "QA" if qa else ""
-        e = "EbNo" if eb_no else ""
-
-        config_name = f"{dataset}{i}{e}{t}{q}Config"
-
-        return getattr(conf, config_name)
+        # to be initialized in setup()
+        self.train: DataLoader = None
+        self.val: DataLoader = None
+        self.test: DataLoader = None
 
     def prepare_data(self) -> None:
-        """Download datasets into self.data_path
-
-        Raises:
-            NotImplementedError: If not implemented by subclass.
         """
-        raise NotImplementedError("prepare_data not implemented")
+        Prepares the dataset by creating new datasets if they do not exist on disk.
+        The datasets are created using the `DatasetCreator` class.
 
-    # create dataset instances
-    # sets self.train, self.val
-    def setup(self, stage: str) -> None:
-        """Set up datasets, self.train and self.val
+        If the dataset already exists on disk, it is loaded back into memory.
+        """
+        # Creates dataset if does not exist on disk or overwrite
+        train_dataset = self.new_dataset_class(
+            dataset_metadata = self.train_metadata,
+        )
+        train_creator = DatasetCreator(
+            dataset = train_dataset,
+            root = self.root,
+            overwrite = False,
+            file_handler = self.file_handler,
+            batch_size = self.create_batch_size,
+            num_workers = self.create_num_workers,
+            train = True,
+        )
+        print(f"Train Dataset: {self.dataset.title()}, Impairment Level {self.train_metadata.impairment_level}, {self.train_metadata.num_samples} samples")
+        train_creator.create()
+
+        val_dataset = self.new_dataset_class(
+            dataset_metadata = self.val_metadata
+        )
+        val_creator = DatasetCreator(
+            dataset = val_dataset,
+            root = self.root,
+            overwrite = False,
+            file_handler = self.file_handler,
+            batch_size = self.create_batch_size,
+            num_workers = self.create_num_workers,
+            train = False,
+        )
+        print(f"Val Dataset: {self.dataset.title()}, Impairment Level {self.val_metadata.impairment_level}, {self.val_metadata.num_samples} samples")
+        val_creator.create()
+
+    def setup(self, stage: str = 'train') -> None:
+        """
+        Sets up the train and validation datasets for the given stage.
 
         Args:
-            stage (str): PyTorch Lightning trainer stage - fit, test, predict.
-
-        Raises:
-            NotImplementedError: If not implemented by subclass.
+            stage (str, optional): The stage of the DataModule, typically 'train' or 'test'. Defaults to 'train'.
         """
-        raise NotImplementedError("setup not implemented")
+        self.train = self.static_dataset_class(
+            root = self.root,
+            impaired = self.impaired,
+            transforms = self.transforms,
+            target_transforms = self.target_transforms,
+            train = True,
+        )
+        self.val = self.static_dataset_class(
+            root = self.root,
+            impaired = self.impaired,
+            transforms = self.transforms,
+            target_transforms = self.target_transforms,
+            train = False
+        )
 
     def train_dataloader(self) -> DataLoader:
-        """Returns train dataloader.
+        """
+        Returns the DataLoader for the training dataset.
 
         Returns:
-            DataLoader: train dataloader
+            DataLoader: A PyTorch DataLoader for the training dataset.
         """
         return DataLoader(
-            dataset=self.train,
-            batch_size=self.batch_size,
-            shuffle=True,
-            num_workers=self.num_workers,
-            collate_fn=self.collate_fn
+            dataset = self.train,
+            batch_size = self.batch_size,
+            shuffle = True,
+            num_workers = self.num_workers,
+            collate_fn = self.collate_fn
         )
-    
+
     def val_dataloader(self) -> DataLoader:
-        """Returns validation dataloader.
+        """
+        Returns the DataLoader for the validation dataset.
 
         Returns:
-            DataLoader: val dataloader
+            DataLoader: A PyTorch DataLoader for the validation dataset.
         """
         return DataLoader(
-            dataset=self.val,
-            batch_size=self.batch_size,
-            shuffle=False,
-            num_workers=self.num_workers,
-            collate_fn=self.collate_fn
+            dataset = self.val,
+            batch_size = self.batch_size,
+            shuffle = False,
+            num_workers = self.num_workers,
+            collate_fn = self.collate_fn
         )
+
+
+### DataModules for creating custom datasets
 
 class NarrowbandDataModule(TorchSigDataModule):
-    """TorchSig Narrowband PyTorch Lightning DataModule
-
-        Attributes:
-            class_list (list): TorchSigNarrowband class list names.
     """
+    DataModule for creating and managing narrowband datasets.
 
-    class_list = list(TorchSigNarrowband._idx_to_name_dict.values())
-    
+    Args:
+        root (str): The root directory where datasets are stored or created.
+        dataset_metadata (NarrowbandMetadata | str | dict): Metadata for the narrowband dataset.
+        num_samples_train (int): The number of training samples.
+        num_samples_val (int, optional): The number of validation samples. Defaults to 10% of training samples if not provided.
+        batch_size (int, optional): The batch size for data loading. Defaults to 1.
+        num_workers (int, optional): The number of worker processes for data loading. Defaults to 1.
+        collate_fn (Callable, optional): A function to collate data into batches.
+        create_batch_size (int, optional): The batch size used during dataset creation. Defaults to 8.
+        create_num_workers (int, optional): The number of workers used during dataset creation. Defaults to 4.
+        file_handler (TorchSigFileHandler, optional): The file handler for managing data storage. Defaults to ZarrFileHandler.
+        transforms (Transform | List[Callable | Transform], optional): A list of transformations to apply to the input data.
+        target_transforms (TargetTransform | List[Callable | TargetTransform], optional): A list of transformations to apply to the target labels.
+    """
+    def __init__(
+        self,
+        root: str,
+        # dataset params
+        dataset_metadata: NarrowbandMetadata | str | dict,
+        num_samples_train: int,
+        num_samples_val: int = None,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: Transform | List[Callable | Transform] = [],
+        target_transforms: TargetTransform | List[Callable | TargetTransform] = [],
+    ):
 
-    def __init__(self, 
-                 root: str, 
-                 impaired: bool, 
-                 qa: bool = True, 
-                 eb_no: bool = False, 
-                 seed: int = 12345,
-                 overlap_prob: Optional[float] = .1,
-                 transform: Optional[Callable] = None,
-                 target_transform: Optional[Callable] = None,
-                 batch_size: int = 4,
-                 num_workers: int = 1,
-                 collate_fn: Optional[Callable] = None
-                ):
-        """TorchSigNarrowband DataModule Init
+        num_samples_val = int(0.10 * num_samples_train) if num_samples_val is None else num_samples_val
 
-        Args:
-            root (str): Dataset root path.
-            impaired (bool): Dataset impairment setting.
-            qa (bool, optional): Generate small dataset sample. Defaults to True.
-            eb_no (bool, optional): Use EbNo config. Defaults to False.
-            seed (int, optional): Dataset generation seed. Defaults to 12345.
-            transform (Optional[Callable], optional): Data transforms. Defaults to None.
-            target_transform (Optional[Callable], optional): Label transforms. Defaults to None.
-            batch_size (int, optional): Dataloader batch size. Defaults to 4.
-            num_workers (int, optional): Dataloader number of workers to use. Defaults to 1.
-            collate_fn (Optional[Callable], optional): Dataloader custom collate function. Defaults to TorchSig collate_fn.
-        """
-        super().__init__(root, "Narrowband", impaired, qa, eb_no, seed, overlap_prob, transform, target_transform, batch_size, num_workers, collate_fn)
+        base = to_dataset_metadata(dataset_metadata)
 
-        self.data_path = f"{self.root}/{self.dataset.lower()}_{self.clean}_"
-        self.train_path = self.data_path + "train"
-        self.val_path = self.data_path + "val"
-
-    def prepare_data(self) -> None:
-        """Download TorchSigNarrowband Dataset
-        """
-        ds_train = ModulationsDataset(
-            level=self.train_config.level,
-            num_samples=self.train_config.num_samples,
-            num_iq_samples=self.train_config.num_iq_samples,
-            use_class_idx=self.train_config.use_class_idx,
-            include_snr=self.train_config.include_snr,
-            eb_no=self.train_config.eb_no,
+        train_metadata = NarrowbandMetadata(
+            num_iq_samples_dataset = base.num_iq_samples_dataset,
+            fft_size = base.fft_size,
+            impairment_level = base.impairment_level,
+            sample_rate = base.sample_rate,
+            num_signals_min = base.num_signals_min,
+            num_signals_distribution = base.num_signals_distribution,
+            snr_db_min = base.snr_db_min,
+            snr_db_max = base.snr_db_max,
+            signal_duration_percent_min = base.signal_duration_percent_min,
+            signal_duration_percent_max = base.signal_duration_percent_max,
+            transforms = base.transforms,
+            target_transforms = base.target_transforms,
+            class_list = base.class_list,
+            class_distribution = base.class_distribution,
+            num_samples = num_samples_train,
         )
-        
-        ds_val = ModulationsDataset(
-            level=self.val_config.level,
-            num_samples=self.val_config.num_samples,
-            num_iq_samples=self.val_config.num_iq_samples,
-            use_class_idx=self.val_config.use_class_idx, 
-            include_snr=self.val_config.include_snr,
-            eb_no=self.val_config.eb_no,
-        )
-    
-        os.makedirs(self.train_path, exist_ok=True)
-        os.makedirs(self.val_path, exist_ok=True)
-    
-        creator_train = DatasetCreator(ds_train, seed=self.seed, path=self.train_path, num_workers = self.num_workers)
-        print(f"Using {self.train_config.__name__} for train.")
-        creator_train.create()
-    
-        creator_val = DatasetCreator(ds_val, seed=self.seed, path=self.val_path, num_workers = self.num_workers)
-        print(f"Using {self.val_config.__name__} for val.")
-        creator_val.create()
 
-    def setup(self, stage: str) -> None:
-        """Set up TorchSigNarrowband train and validation datasets.
-
-        Args:
-            stage (str): PyTorch Lightning trainer stage - fit, test, predict.
-        """
-        self.train = TorchSigNarrowband(
-            self.root,
-            train=True,
-            impaired=self.impaired,
-            transform=self.transform,
-            target_transform=self.target_transform,
-            use_signal_data=True,
+        val_metadata = NarrowbandMetadata(
+            num_iq_samples_dataset = base.num_iq_samples_dataset,
+            fft_size = base.fft_size,
+            impairment_level = base.impairment_level,
+            sample_rate = base.sample_rate,
+            num_signals_min = base.num_signals_min,
+            num_signals_distribution = base.num_signals_distribution,
+            snr_db_min = base.snr_db_min,
+            snr_db_max = base.snr_db_max,
+            signal_duration_percent_min = base.signal_duration_percent_min,
+            signal_duration_percent_max = base.signal_duration_percent_max,
+            transforms = base.transforms,
+            target_transforms = base.target_transforms,
+            class_list = base.class_list,
+            class_distribution = base.class_distribution,
+            num_samples = num_samples_val,
         )
-        
-        self.val = TorchSigNarrowband(
-            self.root,
-            train=False,
-            impaired=self.impaired,
-            transform=self.transform,
-            target_transform=self.target_transform,
-            use_signal_data=True,
+
+        super().__init__(
+            root = root,
+            dataset = 'narrowband',
+            train_metadata = train_metadata,
+            val_metadata = val_metadata,
+            batch_size = batch_size,
+            num_workers = num_workers,
+            collate_fn = collate_fn,
+            create_batch_size = create_batch_size,
+            create_num_workers = create_num_workers,
+            file_handler = file_handler,
+            transforms = transforms,
+            target_transforms = target_transforms,
         )
 
 class WidebandDataModule(TorchSigDataModule):
-    """TorchSigWideband PyTorch Lightning DataModule
+    """
+    DataModule for creating and managing wideband datasets.
 
+    Args:
+        root (str): The root directory where datasets are stored or created.
+        dataset_metadata (WidebandMetadata | str | dict): Metadata for the wideband dataset.
+        num_samples_train (int): The number of training samples.
+        num_samples_val (int, optional): The number of validation samples. Defaults to 10% of training samples if not provided.
+        batch_size (int, optional): The batch size for data loading. Defaults to 1.
+        num_workers (int, optional): The number of worker processes for data loading. Defaults to 1.
+        collate_fn (Callable, optional): A function to collate data into batches.
+        create_batch_size (int, optional): The batch size used during dataset creation. Defaults to 8.
+        create_num_workers (int, optional): The number of workers used during dataset creation. Defaults to 4.
+        file_handler (TorchSigFileHandler, optional): The file handler for managing data storage. Defaults to ZarrFileHandler.
+        transforms (Transform | List[Callable | Transform], optional): A list of transformations to apply to the input data.
+        target_transforms (TargetTransform | List[Callable | TargetTransform], optional): A list of transformations to apply to the target labels.
+    """
+    def __init__(
+        self,
+        root: str,
+        # dataset params
+        dataset_metadata: WidebandMetadata | str | dict,
+        num_samples_train: int,
+        num_samples_val: int = None,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: Transform | List[Callable | Transform]  = [],
+        target_transforms: TargetTransform | List[Callable | TargetTransform] = [],
+    ):
+
+        num_samples_val = int(0.10 * num_samples_train) if num_samples_val is None else num_samples_val
+
+        base = to_dataset_metadata(dataset_metadata)
+
+        train_metadata = WidebandMetadata(
+            num_iq_samples_dataset = base.num_iq_samples_dataset,
+            fft_size = base.num_iq_samples_dataset,
+            impairment_level = base.impairment_level,
+            num_signals_max = base.num_signals_max,
+            sample_rate = base.sample_rate,
+            num_signals_min = base.num_signals_min,
+            num_signals_distribution = base.num_signals_distribution,
+            snr_db_min = base.snr_db_min,
+            snr_db_max = base.snr_db_max,
+            signal_duration_percent_min = base.signal_duration_percent_min,
+            signal_duration_percent_max = base.signal_duration_percent_max,
+            transforms = base.transforms,
+            target_transforms = base.target_transforms,
+            class_list = base.class_list,
+            class_distribution = base.class_distribution,
+            num_samples = num_samples_train,
+        )
+
+        val_metadata = WidebandMetadata(
+            num_iq_samples_dataset = base.num_iq_samples_dataset,
+            fft_size = base.num_iq_samples_dataset,
+            impairment_level = base.impairment_level,
+            num_signals_max = base.num_signals_max,
+            sample_rate = base.sample_rate,
+            num_signals_min = base.num_signals_min,
+            num_signals_distribution = base.num_signals_distribution,
+            snr_db_min = base.snr_db_min,
+            snr_db_max = base.snr_db_max,
+            signal_duration_percent_min = base.signal_duration_percent_min,
+            signal_duration_percent_max = base.signal_duration_percent_max,
+            transforms = base.transforms,
+            target_transforms = base.target_transforms,
+            class_list = base.class_list,
+            class_distribution = base.class_distribution,
+            num_samples = num_samples_val,
+        )
+
+        super().__init__(
+            root = root,
+            dataset = 'wideband',
+            train_metadata = train_metadata,
+            val_metadata = val_metadata,
+            batch_size = batch_size,
+            num_workers = num_workers,
+            collate_fn = collate_fn,
+            create_batch_size = create_batch_size,
+            create_num_workers = create_num_workers,
+            file_handler = file_handler,
+            transforms = transforms,
+            target_transforms = target_transforms,
+        )
+
+
+
+### DataModules for Official Narrowband and Wideband Datasets
+### uses default YAML configs in torchsig/datasets/default_configs
+
+class OfficialTorchSigDataModdule(TorchSigDataModule):
+    """
+    A PyTorch Lightning DataModule for official TorchSignal datasets.
+    
+    This class manages the dataset metadata, configuration, and data loading process 
+    for datasets with official configurations instead of using custom metadata. 
+    It initializes the train and validation metadata based on the dataset type and 
+    impairment level.
+
+    Args:
+        root (str): Root directory where the dataset is stored.
+        dataset (str): Name of the dataset.
+        impaired (bool | int): Defines the impairment level of the dataset.
+        batch_size (int, optional): Batch size for the dataloaders. Default is 1.
+        num_workers (int, optional): Number of workers for data loading. Default is 1.
+        collate_fn (Callable, optional): Function to merge a list of samples into a batch. Default is None.
+        create_batch_size (int, optional): Batch size used during dataset creation. Default is 8.
+        create_num_workers (int, optional): Number of workers used during dataset creation. Default is 4.
+        file_handler (TorchSigFileHandler, optional): File handler used to read/write dataset. Default is ZarrFileHandler.
+        transforms (Transform | List[Callable | Transform], optional): List of transforms applied to dataset. Default is empty list.
+        target_transforms (TargetTransform | List[Callable | TargetTransform], optional): List of transforms applied to targets. Default is empty list.
     """
 
-    def __init__(self, 
-                 root: str, 
-                 impaired: bool, 
-                 qa: bool = True,
-                 seed: int = 12345,
-                 overlap_prob: Optional[float] = None,
-                 fft_size: int = 512,
-                 num_classes: int = 53,
-                 transform: Optional[Callable] = None,
-                 target_transform: Optional[Callable] = None,
-                 batch_size: int = 4,
-                 num_workers: int = 1,
-                 collate_fn: Optional[Callable] = None
-                ):
-        """TorchSigWideband DataModule Init
+    def __init__(
+        self,
+        root: str,
+        dataset: str,
+        impaired: bool | int,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: Transform | List[Callable | Transform]  = [],
+        target_transforms: TargetTransform | List[Callable | TargetTransform] = [],
+    ):
+        # sets train and val metadata
 
-        Args:
-            root (str): Dataset root path.
-            impaired (bool): Dataset impairment setting.
-            qa (bool, optional): Generate small dataset sample. Defaults to True.
-            eb_no (bool, optional): Use EbNo config. Defaults to False.
-            seed (int, optional): Dataset generation seed. Defaults to 12345.
-            overlap_prob (Optional[float], optional): Signal overlap probability. Defaults to None.
-            fft_size (int, optional): Fast Fourier Transform window size. Defaults to 512.
-            num_classes (int, optional): Number of classes of signals. Defaults to 53.
-            transform (Optional[Callable], optional): Data transforms. Defaults to None.
-            target_transform (Optional[Callable], optional): Label transforms. Defaults to None.
-            batch_size (int, optional): Dataloader batch size. Defaults to 4.
-            num_workers (int, optional): Dataloader number of workers to use. Defaults to 1.
-            collate_fn (Optional[Callable], optional): Dataloader custom collate function. Defaults to TorchSig collate_fn.
-        """
-        super().__init__(root, "Wideband", impaired, qa, False, seed, overlap_prob, transform, target_transform, batch_size, num_workers, collate_fn)
-
-        self.overlap_prob = self.train_config.overlap_prob if overlap_prob is None else overlap_prob
-        self.fft_size = fft_size
-        self.num_classes = num_classes
-
-        self.data_path = f"{self.root}/wideband_{self.clean}_"
-        self.train_path = self.data_path + "train"
-        self.val_path = self.data_path + "val"
-
-    def prepare_data(self) -> None:
-        """Download TorchSigWideband
-        """
-        ds_train = WidebandModulationsDataset(
-            level=self.train_config.level,
-            num_samples=self.train_config.num_samples,
-            num_iq_samples=self.train_config.num_iq_samples,
-            seed=self.train_config.seed,
-            overlap_prob=self.overlap_prob
+        train_metadata = get_default_yaml_config(
+            dataset_type = dataset,
+            impairment_level = impaired,
+            train = True
         )
-        
-        ds_val = WidebandModulationsDataset(
-            level=self.val_config.level,
-            num_samples=self.val_config.num_samples,
-            num_iq_samples=self.val_config.num_iq_samples,
-            seed=self.val_config.seed,
-            overlap_prob=self.overlap_prob
-        )
-    
-        os.makedirs(self.train_path, exist_ok=True)
-        os.makedirs(self.val_path, exist_ok=True)
-    
-        creator_train = DatasetCreator(ds_train, seed=self.seed, path=self.train_path, num_workers = self.num_workers)
-        print(f"Using {self.train_config.__name__} for train.")
-        creator_train.create()
-    
-        creator_val = DatasetCreator(ds_val, seed=self.seed, path=self.val_path, num_workers = self.num_workers)
-        print(f"Using {self.val_config.__name__} for val.")
-        creator_val.create()
 
-    def setup(self, stage: str) -> None:
-        """Set up TorchSigWideband train and validation datasets.
-
-        Args:
-            stage (str): PyTorch Lightning trainer stage - fit, test, predict.
-        """
-        self.train = TorchSigWideband(
-            self.root,
-            train=True,
-            impaired=self.impaired,
-            transform=self.transform,
-            target_transform=self.target_transform
+        val_metadata = get_default_yaml_config(
+            dataset_type = dataset,
+            impairment_level = impaired,
+            train = False
         )
-        
-        self.val = TorchSigWideband(
-            self.root,
-            train=False,
-            impaired=self.impaired,
-            transform=self.transform,
-            target_transform=self.target_transform,
+
+        super().__init__(
+            root = root,
+            dataset = dataset,
+            train_metadata = train_metadata,
+            val_metadata = val_metadata,
+            batch_size = batch_size,
+            num_workers = num_workers,
+            collate_fn = collate_fn,
+            create_batch_size = create_batch_size,
+            create_num_workers = create_num_workers,
+            file_handler = file_handler,
+            transforms = transforms,
+            target_transforms = target_transforms,
+        )
+
+
+class OfficialNarrowbandDataModule(OfficialTorchSigDataModdule):
+    """
+    A DataModule for the official Narrowband dataset.
+    
+    This class extends `OfficialTorchSigDataModdule` and sets the dataset type to 'narrowband'.
+    It initializes the necessary parameters for the dataset and loads the train and validation metadata 
+    accordingly.
+
+    Args:
+        root (str): Root directory where the dataset is stored.
+        impaired (bool | int): Defines the impairment level of the dataset.
+        batch_size (int, optional): Batch size for the dataloaders. Default is 1.
+        num_workers (int, optional): Number of workers for data loading. Default is 1.
+        collate_fn (Callable, optional): Function to merge a list of samples into a batch. Default is None.
+        create_batch_size (int, optional): Batch size used during dataset creation. Default is 8.
+        create_num_workers (int, optional): Number of workers used during dataset creation. Default is 4.
+        file_handler (TorchSigFileHandler, optional): File handler used to read/write dataset. Default is ZarrFileHandler.
+        transforms (list, optional): List of transforms applied to dataset. Default is empty list.
+        target_transforms (list, optional): List of transforms applied to targets. Default is empty list.
+    """
+    def __init__(
+        self,
+        root: str,
+        impaired: bool | int,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: list = [],
+        target_transforms: list = [],
+
+    ):
+        # read from yaml or dataset metadata or code inputs
+        super().__init__(
+            root = root,
+            dataset = 'narrowband',
+            impaired = impaired,
+            batch_size = batch_size,
+            num_workers = num_workers,
+            collate_fn = collate_fn,
+            create_batch_size = create_batch_size,
+            create_num_workers = create_num_workers,
+            file_handler = file_handler,
+            transforms = transforms,
+            target_transforms = target_transforms,
+        )
+
+
+class OfficialWidebandDataModule(OfficialTorchSigDataModdule):
+    """
+    A DataModule for the official Wideband dataset.
+    
+    This class extends `OfficialTorchSigDataModdule` and sets the dataset type to 'wideband'.
+    It initializes the necessary parameters for the dataset and loads the train and validation metadata 
+    accordingly.
+
+    Args:
+        root (str): Root directory where the dataset is stored.
+        impaired (bool | int): Defines the impairment level of the dataset.
+        batch_size (int, optional): Batch size for the dataloaders. Default is 1.
+        num_workers (int, optional): Number of workers for data loading. Default is 1.
+        collate_fn (Callable, optional): Function to merge a list of samples into a batch. Default is None.
+        create_batch_size (int, optional): Batch size used during dataset creation. Default is 8.
+        create_num_workers (int, optional): Number of workers used during dataset creation. Default is 4.
+        file_handler (TorchSigFileHandler, optional): File handler used to read/write dataset. Default is ZarrFileHandler.
+        transforms (list, optional): List of transforms applied to dataset. Default is empty list.
+        target_transforms (list, optional): List of transforms applied to targets. Default is empty list.
+    """
+    def __init__(
+        self,
+        root: str,
+        impaired: bool | int,
+        # dataloader params
+        batch_size: int = 1,
+        num_workers: int = 1,
+        collate_fn: Callable = None,
+        # dataset creator params
+        create_batch_size: int = 8,
+        create_num_workers: int = 4,
+        file_handler: TorchSigFileHandler = ZarrFileHandler,
+        # applied after dataset written to disk
+        transforms: list = [],
+        target_transforms: list = [],
+    ):
+        # read from yaml or dataset metadata or code inputs
+        super().__init__(
+            root = root,
+            dataset = 'wideband',
+            impaired = impaired,
+            batch_size = batch_size,
+            num_workers = num_workers,
+            collate_fn = collate_fn,
+            create_batch_size = create_batch_size,
+            create_num_workers = create_num_workers,
+            file_handler = file_handler,
+            transforms = transforms,
+            target_transforms = target_transforms,
         )
