@@ -42,6 +42,7 @@ extensions = [
 
 autosummary_generate = True
 autodoc_member_order = 'bysource'
+autodoc_default_flags = ['members', 'inherited-members']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -64,9 +65,9 @@ copyright = u'2022, torchsig contributors'
 # built documents.
 #
 # The short X.Y version.
-version = '0.6'
+version = '1.0'
 # The full version, including alpha/beta/rc tags.
-release = '0.6.0'
+release = '1.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
