@@ -1,83 +1,57 @@
-from torchsig.transforms.target_transforms import DescToListTuple
-from torchsig.utils.writer import DatasetCreator, DatasetLoader
-from torchsig.datasets.wideband import WidebandModulationsDataset
-from torchsig.datasets import conf
-from torchsig.datasets.signal_classes import torchsig_signals
-from torchsig.transforms.transforms import *
-from torchsig.utils.dataset import collate_fn
-from typing import List
-import click
+"""Wideband generation code for command line
+
+Example:
+To generate Wideband with 10 samples, each sample having 100 IQ samples and 3 signals, and impaired:
+    >>> python generate_wideband.py <path to root> --num_signals=3 --num_samples=10 --num_iq_samples=100 --impaired
+"""
+
+# TorchSig
+from torchsig.datasets.dataset_metadata import WidebandMetadata
+from torchsig.utils.generate import generate
+
+# Third Party
+
+
+# Built-In
 import os
+import argparse
+import math
+import subprocess
+import sys
 
-import numpy as np
+def main():
+    subprocess.check_call([sys.executable, "-m", "pip", "install", "zarr"])
+    from torchsig.utils.generate import generate
+    parser = argparse.ArgumentParser()
 
-modulation_list = torchsig_signals.class_list
+    parser.add_argument("root", type=str, help="Path to generate Wideband dataset.")
+    parser.add_argument("--num_signals", type=int, default=3, help="Number of signals per sample. Defaults to 3.")
+    parser.add_argument("--num_samples", type=int, default=100, help= "Dataset size. Defaults to 100")
+    parser.add_argument("--num_iq_samples", type=int, default=100, help="Nmber of IQ samples per sample. Defaults to 100")
+    parser.add_argument("--fft_size", type=int, default=-1, help="FFT Size. Defaults to sqrt(num_iq_samples)")
+    parser.add_argument("--impaired", action="store_true", help="Generate impaired dataset.")
+    parser.add_argument("--batch_size", type=int, default = 32, help="Batch size. Defaults to 32.")
+    parser.add_argument("--num_workers", type=int, default=os.cpu_count() // 3, help="Number of workers to generate dataset. Defaults to a third of available CPU cores.")
 
-def generate(root: str, configs: List[conf.WidebandConfig], num_workers: int, num_samples_override: int = -1, num_iq_samples_override: int = -1, batch_size: int = 32):
-    for config in configs:
-        num_samples = config.num_samples if num_samples_override <=0 else num_samples_override
-        num_iq_samples = config.num_iq_samples if num_iq_samples_override <= 0 else num_iq_samples_override
-        # batch_size = 32 if num_workers > config.num_samples else int(np.min((num_samples // num_workers, 32)))
-        prefetch_factor = None if num_workers <= 1 else 4
-        print(f'batch_size -> {batch_size} num_samples -> {num_samples}, config -> {config}')
+    # Parse the arguments
+    args = parser.parse_args()
 
-        wideband_ds = WidebandModulationsDataset(
-            level=config.level,
-            num_iq_samples=num_iq_samples,
-            num_samples=num_samples,
-            modulation_list=modulation_list,
-            # target_transform=DescToListTuple(),
-            seed=config.seed,
-            overlap_prob=config.overlap_prob
-        )
+    # create wideband metadata
+    dataset_metadata = WidebandMetadata(
+        num_signals_max=args.num_signals,
+        num_samples = args.num_samples,
+        num_iq_samples_dataset=args.num_iq_samples,
+        fft_size= int(math.sqrt(args.num_iq_samples)) if args.fft_size == -1 else args.fft_size,
+        impairment_level= 2 if args.impaired else 0,
+    )
 
-        dataset_loader = DatasetLoader(wideband_ds, seed=12345678, collate_fn=collate_fn, num_workers=num_workers, batch_size=batch_size, prefetch_factor=prefetch_factor)
-        creator = DatasetCreator(wideband_ds, seed=12345678, num_workers=num_workers, path=os.path.join(root, config.name), loader=dataset_loader,)
-        creator.create()
-
-
-@click.command()
-@click.option("--root", default="wideband", help="Path to generate wideband datasets")
-@click.option("--all", is_flag=True, default=False, help="Generate all versions of wideband_ dataset.")
-@click.option("--qa", is_flag=True, default=False, help="Generate only QA versions of wideband dataset.")
-@click.option("--num-iq-samples", "num_iq_samples", default=-1, help="Override number of iq samples in wideband dataset.")
-@click.option("--batch-size", "batch_size", default=32, help="Override batch size.")
-@click.option("--num-samples", default=-1, help="Override for number of dataset samples.")
-@click.option("--impaired", is_flag=True, default=False, help="Generate impaired dataset. Ignored if --all (default)",)
-@click.option("--num-workers", "num_workers", default=os.cpu_count() // 2, help="Define number of workers for both DatasetLoader and DatasetCreator")
-def main(root: str, all: bool, qa: bool, impaired: bool, num_workers: int, num_samples: int, num_iq_samples: int, batch_size: int):
-    os.makedirs(root, exist_ok=True)
-
-    configs = [
-        conf.WidebandCleanTrainConfig,
-        conf.WidebandCleanValConfig,
-        conf.WidebandImpairedTrainConfig,
-        conf.WidebandImpairedValConfig,
-        conf.WidebandCleanTrainQAConfig,
-        conf.WidebandCleanValQAConfig,
-        conf.WidebandImpairedTrainQAConfig,
-        conf.WidebandImpairedValQAConfig,
-    ]
-
-    impaired_configs = []
-    impaired_configs.extend(configs[2:4])
-    impaired_configs.extend(configs[-2:])
-    
-    if all:
-        generate(root, configs, num_workers, num_samples, num_iq_samples, batch_size)
-        return
-    
-    elif qa:
-        generate(root, configs[-4:], num_workers, num_samples, num_iq_samples, batch_size)
-        return
-
-    elif impaired:
-        generate(root, impaired_configs, num_workers, num_samples, num_iq_samples, batch_size)
-        return
-
-    else:
-        generate(root, configs[:2], num_workers, num_samples, num_iq_samples, batch_size)
-
+    generate(
+        root=args.root,
+        dataset_metadata=dataset_metadata,
+        batch_size=args.batch_size,
+        num_workers=args.num_workers,
+    )
 
 if __name__ == "__main__":
     main()
+
